Add script to delete all stripe data.
since the stripe data is closely linked to the database, during testing, this script can be used in
tandem with deleting all data in the db along with data in stripe.

Usage: `npm run delete-stripe-data`

The script will use the key set in the `STRIPE_API_KEY` env var.
Alternatively, the `STRIPE_API_KEY` can also be set in the `<rootDir>.env`